requirements.txt
flask
flask_restful
requests
pyzbar
pillow